refactoring certificate_signer to extract sign_certificate as function
